+++ package/src/renderer/components/shaders/Turbulence.tsx
@@ -1,32 +1,10 @@
 import React from "react";
 
-import { createDeclaration } from "../../nodes/Declaration";
-import type { AnimatedProps } from "../../processors/Animations/Animations";
+import type { TurbulenceProps } from "../../../dom/types";
+import type { SkiaProps } from "../../processors/Animations/Animations";
 
-export interface TurbulenceProps {
-  freqX: number;
-  freqY: number;
-  octaves: number;
-  seed: number;
-  tileWidth: number;
-  tileHeight: number;
-}
-
-const onDeclare = createDeclaration<TurbulenceProps>(
-  ({ freqX, freqY, octaves, seed, tileWidth, tileHeight }, _, { Skia }) => {
-    return Skia.Shader.MakeTurbulence(
-      freqX,
-      freqY,
-      octaves,
-      seed,
-      tileWidth,
-      tileHeight
-    );
-  }
-);
-
-export const Turbulence = (props: AnimatedProps<TurbulenceProps>) => {
-  return <skDeclaration onDeclare={onDeclare} {...props} />;
+export const Turbulence = (props: SkiaProps<TurbulenceProps>) => {
+  return <skTurbulence {...props} />;
 };
 
 Turbulence.defaultProps = {
